chore: rename routes
